requirements.txt
Django 
djangorestframework 
django-rest-auth
django-heroku
python-dateutil
psycopg2-binary
gunicorn
cryptography
django-extensions
djangorestframework-jwt
PyJWT